solana_sysvar
Module recent_blockhashes

Information about recent blocks and their fee calculators.

The recent blockhashes sysvar provides access to the RecentBlockhashes, which contains recent blockhahes and their FeeCalculators.

RecentBlockhashes does not implement Sysvar::get.

This sysvar is deprecated and should not be used. Transaction fees should be determined with the getFeeForMessage RPC method. For additional context see the Comprehensive Compute Fees proposal.

See also the Solana documentation on the recent blockhashes sysvar.

Structs

EntryDeprecated

IntoIterSorted

IterItemDeprecated

RecentBlockhashesDeprecated

Contains recent block hashes and fee calculators.

Constants

ID

The const program ID.

MAX_ENTRIESDeprecated

Functions

check_id

Returns true if given pubkey is the program ID.

id

Returns the program ID.