solana_sysvar::slot_hashes

Struct SlotHashesSysvar

pub struct SlotHashesSysvar;

๐Deprecated since 2.1.0: Please use PodSlotHashes instead

API for querying the SlotHashes sysvar.

Implementations

impl SlotHashesSysvar

pub fn get(slot: &Slot) -> Result<Option<Hash>, ProgramError>

Get a value from the sysvar entries by its key. Returns None if the key is not found.

pub fn position(slot: &Slot) -> Result<Option<usize>, ProgramError>

Get the position of an entry in the sysvar by its key. Returns None if the key is not found.