solana_sysvar::slot_hashes

Trait SysvarId

pub trait SysvarId {
    // Required methods
    fn id() -> Pubkey;
    fn check_id(pubkey: &Pubkey) -> bool;
}

A type that holds sysvar data and has an associated sysvar Pubkey.

Required Methods

fn id() -> Pubkey

The Pubkey of the sysvar.

fn check_id(pubkey: &Pubkey) -> bool

Returns true if the given pubkey is the program ID.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.