solana_sysvar

Module fees

Current cluster fees.

The fees sysvar provides access to the Fees type, which contains the current FeeCalculator.

Fees implements Sysvar::get and can be loaded efficiently without passing the sysvar account ID to the program.

This sysvar is deprecated and will not be available in the future. Transaction fees should be determined with the getFeeForMessage RPC method. For additional context see the Comprehensive Compute Fees proposal.

See also the Solana documentation on the fees sysvar.

Structs

FeesDeprecated

Transaction fees.

Constants

ID

The const program ID.

Functions

check_id

Returns true if given pubkey is the program ID.

id

Returns the program ID.